+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/ShelfFragment.kt
@@ -35,6 +35,7 @@ import au.com.shiftyjelly.pocketcasts.ui.helper.ColorUtils
 import au.com.shiftyjelly.pocketcasts.ui.helper.FragmentHostListener
 import au.com.shiftyjelly.pocketcasts.ui.theme.ThemeColor
 import au.com.shiftyjelly.pocketcasts.utils.extensions.dpToPx
+import au.com.shiftyjelly.pocketcasts.utils.log.LogBuffer
 import au.com.shiftyjelly.pocketcasts.views.extensions.setRippleBackground
 import au.com.shiftyjelly.pocketcasts.views.fragments.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
@@ -163,8 +164,11 @@ class ShelfFragment : BaseFragment(), ShelfTouchCallback.ItemTouchHelperAdapter 
 
     private fun trackShelfItemMovedEvent(position: Int) {
         dragStartPosition?.let {
-            val title = (items[position] as? ShelfItem)?.analyticsValue
-                ?: AnalyticsProp.Value.UNKNOWN
+            val title = try {
+                (items[position] as? ShelfItem)?.analyticsValue
+            } catch (e: ArrayIndexOutOfBoundsException) {
+                LogBuffer.i(LogBuffer.TAG_INVALID_STATE, "Error getting title for position $position in ShelfFragment", e)
+            } ?: AnalyticsProp.Value.UNKNOWN
             val movedFrom = sectionTitleAt(it)
             val movedTo = sectionTitleAt(position)
             val newPosition = if (movedTo == AnalyticsProp.Value.SHELF) {